requirements.txt
datasets
fastapi
loguru
numpy
Pillow
pre-commitpydantic
python-multipart
pytorch-lightning
torchmetrics 
transformers
uvicorn[standard]
